Factor out element similarity state.
